+++ webpack.config.base.js
@@ -24,7 +24,7 @@ export default {
 
   output: {
     path: path.join(__dirname, 'app'),
-    filename: 'bundle.js',
+    filename: 'renderer.dev.js',
     // https://github.com/webpack/webpack/issues/1114
     libraryTarget: 'commonjs2'
   },
@@ -41,6 +41,10 @@ export default {
   },
 
   plugins: [
+    new webpack.DefinePlugin({
+      'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV || 'production')
+    }),
+
     new webpack.NamedModulesPlugin(),
   ],
 };